backend/pyproject.toml
[project]
name = "f1-dashboard-backend"
version = "0.1.0"
description = "F1 Dashboard FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "fastf1>=3.5.3",
    "pandas>=2.3.0",
    "numpy>=2.2.0",
    "python-dotenv>=1.1.0",
    "pydantic-settings>=2.0.0",
    "sse-starlette>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.23.0",
    "anyio[trio]>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
